// src/app/models/products.ts
export interface Productos {
    id:             string;
    nombre:         string;
    peso:           string;
    descriptcion:   string;
    precioActual:   string;
    precioAnterior: string;
    categoria:      string;
    animal:         string;
    tipoProducto:   string;
    esNuevo:        boolean;
    marca:          string;
    stock:          string;
    srcImg:         string;
    fecha:          string;
}